Refactorized the logi of change the status of the employee and send to the service.
